Remove the `times` package
The thesis has been written without the `times` package, so every
paragraph were formatted according to the default font. Since the default
font is slighly wider, most of the paragraphs looks like nice block of
text with it. With the `times` font, it changes the formatting and some
last lines only have two or three words on it (I try to avoid having
last lines with only two or three words).